+++ pkg/layers/layer.go
@@ -1,4 +1,4 @@
-package guts
+package layers
 
 import (
 	"gonum.org/v1/gonum/mat"
