+++ pkg/loss/cost_function.go
@@ -1,4 +1,4 @@
-package guts
+package loss
 
 import (
 	"gonum.org/v1/gonum/mat"
