+++ pkg/loss/mse.go
@@ -1,4 +1,4 @@
-package guts
+package loss
 
 import (
 	"fmt"
